requirements.txt
asgiref==3.7.2
attrs==23.1.0
Babel==2.12.1
Django==4.2.4
django-money==3.2.0
djangorestframework==3.14.0
drf-spectacular==0.26.4
drf-spectacular-sidecar==2023.8.1
Faker==19.3.0
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
Pillow==10.0.0
postgres==4.0
psycopg2-binary==2.9.7
psycopg2-pool==1.1
py-moneyed==3.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.9.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
